Set users permission

+++ app/Http/Controllers/Admin/UserController.php
@@ -75,10 +75,25 @@ class UserController extends Controller
 				})
 				->editColumn('Actions', function ($row) {
 					$buttons = '';
+
+					$user = User::with(['roles', 'roles.permissions'])->where('id',Auth::user()->id)->first();
+
+					$edit_permission = array_filter($user->roles[0]['permissions']->toArray(), function ($var) {
+						return ($var['name'] == 'user-edit');
+					});
+
+					$delete_permission = array_filter($user->roles[0]['permissions']->toArray(), function ($var) {
+						return ($var['name'] == 'user-delete');
+					});
+
+					if(count($edit_permission) > 0) {
 						$buttons =  $buttons . '<a href="'.route('admin.user.edit',$row->id).'"><i role="button" title="Edit" data-id="' . $row->id . '"  class="fs-22 py-2 mx-2 fa-pencil pointer fa  " type="button"></i></a>';
-		
+					}
+
+					if(count($delete_permission) > 0) {
 						$buttons =  $buttons . '<i role="button" title="Delete" data-id="' . $row->id . '" onclick=deleteUser(this) class="fa-trash pointer fa fs-22 py-2 mx-2 text-danger" type="button"></i>';
-					
+					}
+
 					return $buttons;
 				})
 				->rawColumns(['Actions','status','email'])
